Split movement and rotation for players into individual scripts
