+++ Assets/Scripts/Laser.cs
@@ -10,12 +10,18 @@ public class Laser : MonoBehaviour
     // Start is called before the first frame update
     void Start()
     {
-        
+        lineRenderer = GetComponent<LineRenderer>();
+        lineRenderer.enabled = true;
+        lineRenderer.useWorldSpace = true;
     }
 
     // Update is called once per frame
     void Update()
     {
-        
+        RaycastHit2D hit = Physics2D.Raycast(transform.position, transform.up);
+        Debug.DrawLine(transform.position, hit.point);
+        lineRenderer.SetPosition(0, transform.position);
+        lineRenderer.SetPosition(1, LaserHit.position);
+
     }
 }
