+++ Assets/Scripts/Player/player_rotation.cs
@@ -1,42 +1,22 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-using UnityEngine.UIElements;
 
-public class player_controls : MonoBehaviour
+public class player_rotation : MonoBehaviour
 {
     // Create private rigid body for player
     private Rigidbody2D player_rb;
 
-    // Set player's movement speed
-    [SerializeField]
-    float movementSpeed = 3f;
-
-    // Player movement
-    private Vector2 playerMovement;
-
-    private void Start()
+    // Start is called before the first frame update
+    void Start()
     {
         // Get player's rigid body
         player_rb = GetComponent<Rigidbody2D>();
     }
 
     // Update is called once per frame
-    private void Update()
+    void Update()
     {
-
-        // forward and backward movement
-        float vertical = Input.GetAxisRaw("Vertical");
-        // left and right movement
-        float horizontal = Input.GetAxisRaw("Horizontal");
-
-        // Get the new player movement vector value
-        playerMovement = new Vector2(horizontal, vertical);
-
-        playerMovement.Normalize();
-
-        // This portion of the code rotates the player
-
         // Get current mouse position 
         Vector3 mousePos = Camera.main.ScreenToWorldPoint(Input.mousePosition);
 
@@ -48,11 +28,5 @@ public class player_controls : MonoBehaviour
 
         // Rotate player
         transform.rotation = Quaternion.Euler(new Vector3(0f, 0f, rotationAngle));
-    }
-
-    private void FixedUpdate()
-    {
-        // Move player
-        player_rb.velocity = playerMovement * movementSpeed;
     }
 }
