+++ Assets/Scripts/Player/player_rotation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8df7ce2e1d5cb2341a80c3784ae62119
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
